requirements.txt
pymongo>=4.0
python-dotenv